chore: move of NewPool was too aggressive, moving back to pg driver package
